Fix pointer to interger conversion errors

+++ FlycutEngine/FlycutStore.h
@@ -52,7 +52,7 @@
 -(NSArray *) previousIndexes:(int)howMany containing:(NSString*)search; // This method is in newest-first order.
 
 // Add a clipping
--(void) addClipping:(NSString *)clipping ofType:(NSString *)type fromAppLocalizedName:(NSString *)appLocalizedName fromAppBundleURL:(NSString *)bundleURL atTimestamp:(int) timestamp;
+-(void) addClipping:(NSString *)clipping ofType:(NSString *)type fromAppLocalizedName:(NSString *)appLocalizedName fromAppBundleURL:(NSString *)bundleURL atTimestamp:(NSInteger) timestamp;
 -(void) addClipping:(FlycutClipping*) clipping;
 
 // Delete a clipping
